Mijn tijdelijke DbTablesRepository verwijdert en toegevoegd aan de nieuwe TablesRepository.
