+++ ChapeauApp/Views/Orders/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@model ChapeauApp.Models.ViewModels.OrdersViewModel
-
+@*Deze View verwijderen*@
 @{
     ViewData["Title"] = "View TableOrders";
 }
